refactor: Move LockDeviceDetails to module
